WIP: Hell of a refactoring
