+++ parseText/index.js
@@ -1,0 +1,5 @@
+const parseExpression = require('./parseExpression')
+
+const parseText = string => parseExpression(string.toLowerCase())
+
+module.exports = parseText
